requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cssutils==2.4.0
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.4
django-cors-headers==3.5.0
django-countries==7.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-dotenv==1.4.2
django-extensions==3.1.3
django-filter==2.4.0
django-inlinecss==0.3.0
django-money==1.2.1
django-phone-field==1.8.1
django-phonenumber-field==6.1.0
django-smart-selects==1.3.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-schema-adapter==2.0.7
entrypoints==0.4
et-xmlfile==1.1.0
fastjsonschema==2.15.3
future==0.18.2
fuzzywuzzy==0.18.0
heroku==0.1.4
idna==3.3
Inflector==3.0.1
ipykernel==5.5.6
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
MarkupSafe==2.1.1
mistune==0.8.4
mysqlclient==2.1.1
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.3
oauthlib==3.2.0
openpyxl==3.0.7
packaging==21.3
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.3
phonenumbers==8.12.49
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2==2.9.1
py-moneyed==0.8.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.3.0
pynliner==0.8.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pywin32==304
pywinpty==2.0.5
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
requests==2.27.1
requests-oauthlib==1.3.1
schedule==1.0.0
Send2Trash==1.8.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
terminado==0.13.3
tinycss2==1.1.1
toml==0.10.2
tornado==6.1
traitlets==5.2.0
twilio==7.9.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
xlrd==2.0.1
XlsxWriter==1.4.3
